+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## v17.2.2.14 / 2017 Sep 05
+* **Fix** - Forcing `AsyncEventReporter` to use `WebRequest` under the hood as `HttpClient` will cause deadlocks in 
+systems with asynchronous entry points.
+
+```csharp
+[GuaranteedRate.Sextant "17.2.2.14"]
+```
+
 ## v17.2.2.13 / 2017 Sep 05
 * **Fix** - Do not lock the log statement in `Logger.cs`